moved bookmark routes to userrouting, changed bookmark button to accept props instead of fetching on its own
